+++ tsconfig.json
@@ -9,6 +9,6 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true
   },
-  "include": ["colors.ts", "compile-json.ts", "./charts/**/*"],
+  "include": ["colors.ts", "compile-json.ts", "./chart/**/*"],
   "exclude": ["node_modules"]
 }
